requirements.txt
# Tinbot/requirements.txt
openai>=1.0.0
python-dotenv
pillow
pyautogui
pyperclip
DrissionPage>=4.0.0
rich